% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_integrity.R
\name{add_time_steps}
\alias{add_time_steps}
\title{Add timestep columns to a dataframe}
\usage{
add_time_steps(
  .data,
  date_column = "survey_datetime",
  species = c("chinook", "coho")
)
}
\arguments{
\item{.data}{dataframe}

\item{date_column}{Column of the datetime}

\item{species}{Species, either chinook or coho}
}
\description{
Adds a time step column to a dataframe
}
\examples{
\dontrun{df |> add_time_steps()}

}
