% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aerial.R
\name{aerial_estimator}
\alias{aerial_estimator}
\title{Aerial Estimator Table}
\usage{
aerial_estimator(catch_area, start_date, end_date)
}
\arguments{
\item{catch_area}{Valid catch area code}

\item{start_date}{Start date of search}

\item{end_date}{End date of search}
}
\description{
Finds in-sample boats during the flight and calculates various numbers from it
}
\examples{
\dontrun{df <- aerial_estimator('07', '2021-07-01', '2021-07-07')}
}
\seealso{
Other Aerial Effort Surveys: 
\code{\link{aerial_estimator_stats}()},
\code{\link{aerial_pull}()}
}
\concept{Aerial Effort Surveys}
