% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aerial.R
\name{aerial_estimator_stats}
\alias{aerial_estimator_stats}
\title{Aerial Estimator Table}
\usage{
aerial_estimator_stats(.data, bad_flights = FALSE)
}
\arguments{
\item{.data}{Dataframe}

\item{bad_flights}{Throws out flights where there were more boats sampled at the dock than see from the air}
}
\description{
Calculaties the adjusted proportion and variance of the aerial estimator table
}
\examples{
\dontrun{df <- aerial_estimator('07', '2021-07-01', '2021-07-07')
df |>
      aerial_estimator_stats()
      }
}
\seealso{
Other Aerial Effort Surveys: 
\code{\link{aerial_estimator}()},
\code{\link{aerial_pull}()}
}
\concept{Aerial Effort Surveys}
