% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aerial.R
\name{aerial_pull}
\alias{aerial_pull}
\title{Aerial data pull}
\usage{
aerial_pull(
  catch_area,
  start_date,
  end_date,
  fresh = FALSE,
  check_fresh = TRUE
)
}
\arguments{
\item{catch_area}{Valid catch area code}

\item{start_date}{Start date of search}

\item{end_date}{End date of search}

\item{fresh}{Boolean. Ensures data is updated}

\item{check_fresh}{Boolean. Checks whether data is current}
}
\description{
Pulls the aerial data
}
\examples{
\dontrun{df <- aerial_pull('07', '2021-07-01', '2021-07-07')}
}
\seealso{
Other Aerial Effort Surveys: 
\code{\link{aerial_estimator}()},
\code{\link{aerial_estimator_stats}()}
}
\concept{Aerial Effort Surveys}
