% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boatsurveys.R
\name{bs_map}
\alias{bs_map}
\title{Effort map}
\usage{
bs_map(.data, blur = 25, max = 0.05, radius = 10)
}
\arguments{
\item{.data}{Dataframe}

\item{blur}{Inherited argument from leaflet}

\item{max}{Inherited argument from leaflet}

\item{radius}{Inherited argument from leaflet}
}
\description{
Provides a heat map of angling effort
}
\examples{
\dontrun{bs_data |> bs_map()}
}
\seealso{
Other Boat Effort Surveys: 
\code{\link{bs_pull}()},
\code{\link{bs_site_weights}()},
\code{\link{bs_site_weights_in}()},
\code{\link{bs_site_weights_out}()}
}
\concept{Boat Effort Surveys}
