% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boatsurveys.R
\name{bs_pull}
\alias{bs_pull}
\title{Downloads test fishing data from pssp}
\usage{
bs_pull(
  catch_area,
  start_date,
  end_date,
  fresh = FALSE,
  check_fresh = TRUE,
  dsn = "pssp_prod64"
)
}
\arguments{
\item{catch_area}{Valid catch area code}

\item{start_date}{Start date of search}

\item{end_date}{End date of search}

\item{fresh}{Boolean. Ensures data is updated}

\item{check_fresh}{Boolean. Checks whether data is current}

\item{dsn}{DSN for the database connection. Default 'pssp_prod64'}
}
\description{
Queries RMIS for release infomation, defaults to tagged releases
}
\examples{
\dontrun{bs_data <- bs_pull('10', '2021-11-01', '2021-12-04')}
}
\seealso{
Other Boat Effort Surveys: 
\code{\link{bs_map}()},
\code{\link{bs_site_weights}()},
\code{\link{bs_site_weights_in}()},
\code{\link{bs_site_weights_out}()}
}
\concept{Boat Effort Surveys}
