% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boatsurveys.R
\name{bs_site_weights}
\alias{bs_site_weights}
\title{Calculates site weights}
\usage{
bs_site_weights(.data, sample_type = c("in_sample", "out_sample"))
}
\arguments{
\item{.data}{Dataframe}

\item{sample_type}{Return in sample site weights or out of sample site weights}
}
\description{
Calculates site weights
}
\examples{
\dontrun{bs_data |> bs_site_weights(sample_type = 'in_sample')}
\dontrun{bs_data |> bs_site_weights(sample_type = 'out_sample')}
}
\seealso{
Other Boat Effort Surveys: 
\code{\link{bs_map}()},
\code{\link{bs_pull}()},
\code{\link{bs_site_weights_in}()},
\code{\link{bs_site_weights_out}()}
}
\concept{Boat Effort Surveys}
