% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boatsurveys.R
\name{bs_site_weights_in}
\alias{bs_site_weights_in}
\title{Calculates in-sample site weights}
\usage{
bs_site_weights_in(.data)
}
\arguments{
\item{.data}{Dataframe}
}
\description{
Calculates in-sample site weights
}
\examples{
\dontrun{bs_data |> bs_site_weights_in()}
}
\seealso{
Other Boat Effort Surveys: 
\code{\link{bs_map}()},
\code{\link{bs_pull}()},
\code{\link{bs_site_weights}()},
\code{\link{bs_site_weights_out}()}
}
\concept{Boat Effort Surveys}
