% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boatsurveys.R
\name{bs_site_weights_out}
\alias{bs_site_weights_out}
\title{Calculates out-of-sample site weights}
\usage{
bs_site_weights_out(.data)
}
\arguments{
\item{.data}{Dataframe}
}
\description{
Calculates out-of-sample site weights
}
\examples{
\dontrun{bs_data |> bs_site_weights_out()}
}
\seealso{
Other Boat Effort Surveys: 
\code{\link{bs_map}()},
\code{\link{bs_pull}()},
\code{\link{bs_site_weights}()},
\code{\link{bs_site_weights_in}()}
}
\concept{Boat Effort Surveys}
