% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_integrity.R
\name{calculate_total_length}
\alias{calculate_total_length}
\title{Total length calculation for fork length}
\usage{
calculate_total_length(
  .data,
  fork_length_column = "fork_length",
  size_class = TRUE,
  minimum_size_limit_inches = 22
)
}
\arguments{
\item{.data}{Dataframe}

\item{fork_length_column}{Column name of fork lengths provided}

\item{size_class}{Adds size class column (sublegal and legal)}

\item{minimum_size_limit_inches}{Size limit (total length) of fishery in inches}
}
\description{
Adds StatWeek to Dataframe
}
\examples{
\dontrun{
df |> calculate_total_length()
}

}
