% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_integrity.R
\name{check_freshness}
\alias{check_freshness}
\title{Checks if data is current}
\usage{
check_freshness(days = 3)
}
\arguments{
\item{days}{The number of days to check with the last update of the materialized views}
}
\description{
Checks if data is current, if it is then updates, if not passes through
}
\examples{
\dontrun{check_freshness()}
}
