% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockside.R
\name{ds_encounters}
\alias{ds_encounters}
\title{Pulls the dockside encounters}
\usage{
ds_encounters(
  catch_area,
  start_date,
  end_date,
  pretty = TRUE,
  bias_corrected = TRUE,
  minimum_size_limit_inches = 22
)
}
\arguments{
\item{catch_area}{Valid catch area code}

\item{start_date}{Start date of search}

\item{end_date}{End date of search}

\item{pretty}{If TRUE will return a summarized dataframe, if FALSE will return a list to be used in other calculations}

\item{bias_corrected}{Will return proportions with bias correction}

\item{minimum_size_limit_inches}{Minimum size limit for fishery}
}
\description{
Pulls Dockside Encounters
}
\examples{
\dontrun{dockside_encounters <- ds_encounters('10', '2021-11-01', '2021-12-04')}
}
\seealso{
Other Dockside Creel Surveys: 
\code{\link{ds_bio_pull}()},
\code{\link{ds_filter_insample}()},
\code{\link{ds_filter_insample_aerial}()},
\code{\link{ds_filter_insample_murthy}()},
\code{\link{ds_historical_cpue_chart}()},
\code{\link{ds_pull_sampling_effort}()},
\code{\link{ds_releases_pull}()},
\code{\link{ds_salmon_summary}()},
\code{\link{ds_sampling_effort_sampler_summary}()},
\code{\link{ds_sampling_effort_summary}()},
\code{\link{identify_in_sample}()}
}
\concept{Dockside Creel Surveys}
