% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_sample_filtering.R
\name{ds_filter_insample_murthy}
\alias{ds_filter_insample_murthy}
\title{In-Sample filter (Murthy)}
\usage{
ds_filter_insample_murthy(.data)
}
\arguments{
\item{.data}{Dataframe}
}
\description{
Filters to in sample locations with a Murthy study design
}
\examples{
\dontrun{df |> ds_filter_insample_murthy()}
}
\seealso{
Other Dockside Creel Surveys: 
\code{\link{ds_bio_pull}()},
\code{\link{ds_encounters}()},
\code{\link{ds_filter_insample}()},
\code{\link{ds_filter_insample_aerial}()},
\code{\link{ds_historical_cpue_chart}()},
\code{\link{ds_pull_sampling_effort}()},
\code{\link{ds_releases_pull}()},
\code{\link{ds_salmon_summary}()},
\code{\link{ds_sampling_effort_sampler_summary}()},
\code{\link{ds_sampling_effort_summary}()},
\code{\link{identify_in_sample}()},
\code{\link{marine_pull}()}
}
\concept{Dockside Creel Surveys}
