% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockside.R
\name{ds_historical_cpue_chart}
\alias{ds_historical_cpue_chart}
\title{Makes a chart comparing current year CPUE vs previous years}
\usage{
ds_historical_cpue_chart(
  catch_area,
  start_date,
  end_date,
  species = c("chinook", "coho", "pink"),
  smoothing_method = "loess",
  clipped_only = T
)
}
\arguments{
\item{catch_area}{Valid catch area code}

\item{start_date}{Start date of search}

\item{end_date}{End date of search}

\item{species}{Species to look for - chinook, pink, coho}

\item{smoothing_method}{Default loess. Takes method arguments from \code{geom_smooth}}

\item{clipped_only}{Will use only ad clipped fish}
}
\description{
Makes a chart comparing current year CPUE vs previous years
}
\examples{
\dontrun{ds_historical_cpue_chart('10', '2021-11-01', '2021-12-04')}
\dontrun{ds_historical_cpue_chart('10', '2021-11-01', '2021-12-04') + ggtitle('New Title')}
}
\seealso{
Other Dockside Creel Surveys: 
\code{\link{ds_bio_pull}()},
\code{\link{ds_encounters}()},
\code{\link{ds_filter_insample}()},
\code{\link{ds_filter_insample_aerial}()},
\code{\link{ds_filter_insample_murthy}()},
\code{\link{ds_pull_sampling_effort}()},
\code{\link{ds_releases_pull}()},
\code{\link{ds_salmon_summary}()},
\code{\link{ds_sampling_effort_sampler_summary}()},
\code{\link{ds_sampling_effort_summary}()},
\code{\link{identify_in_sample}()},
\code{\link{marine_pull}()}
}
\concept{Dockside Creel Surveys}
