% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockside.R
\name{ds_pull_sampling_effort}
\alias{ds_pull_sampling_effort}
\title{Get sampling effort data}
\usage{
ds_pull_sampling_effort(marine_area, start_date, end_date)
}
\arguments{
\item{marine_area}{Valid marine area code}

\item{start_date}{Start date of search}

\item{end_date}{End date of search}
}
\description{
Gets effort data for our dockside sampling
}
\examples{
\dontrun{df <- ds_pull_sampling_effort('10', '2021-01-01', '2021-12-31')}
}
\seealso{
Other Dockside Creel Surveys: 
\code{\link{ds_bio_pull}()},
\code{\link{ds_encounters}()},
\code{\link{ds_filter_insample}()},
\code{\link{ds_filter_insample_aerial}()},
\code{\link{ds_filter_insample_murthy}()},
\code{\link{ds_historical_cpue_chart}()},
\code{\link{ds_releases_pull}()},
\code{\link{ds_salmon_summary}()},
\code{\link{ds_sampling_effort_sampler_summary}()},
\code{\link{ds_sampling_effort_summary}()},
\code{\link{identify_in_sample}()},
\code{\link{marine_pull}()}
}
\concept{Dockside Creel Surveys}
