% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockside.R
\name{ds_releases_pull}
\alias{ds_releases_pull}
\title{Pulls release information for chinook and coho with legal size status}
\usage{
ds_releases_pull(
  catch_area,
  start_date,
  end_date,
  int_dates = FALSE,
  fill_zero = FALSE,
  check_fresh = TRUE,
  fresh = FALSE
)
}
\arguments{
\item{catch_area}{Valid catch area code}

\item{start_date}{Start date of search}

\item{end_date}{End date of search}

\item{int_dates}{Returns year, month, day integer columns instead of a date field}

\item{fill_zero}{Boolean. Fills NA's with zeros.}

\item{check_fresh}{Boolean. Checks whether data is current}

\item{fresh}{Boolean. Ensures data is updated}
}
\description{
Pulls release information for chinook and coho with legal size status
}
\examples{
\dontrun{ds_rel_data <- ds_releases_pull('10', '2021-11-01', '2021-12-04')}
\dontrun{ds_rel_data <- ds_releases_pull(c('10', '09'), '2021-11-01', '2021-12-04')}
}
\seealso{
Other Dockside Creel Surveys: 
\code{\link{ds_bio_pull}()},
\code{\link{ds_encounters}()},
\code{\link{ds_filter_insample}()},
\code{\link{ds_filter_insample_aerial}()},
\code{\link{ds_filter_insample_murthy}()},
\code{\link{ds_historical_cpue_chart}()},
\code{\link{ds_pull_sampling_effort}()},
\code{\link{ds_salmon_summary}()},
\code{\link{ds_sampling_effort_sampler_summary}()},
\code{\link{ds_sampling_effort_summary}()},
\code{\link{identify_in_sample}()}
}
\concept{Dockside Creel Surveys}
