% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockside.R
\name{ds_salmon_summary}
\alias{ds_salmon_summary}
\title{Downloads test fishing data from pssp}
\usage{
ds_salmon_summary(
  catch_area,
  start_date,
  end_date,
  int_dates = FALSE,
  fill_zero = FALSE,
  fresh = FALSE,
  check_fresh = TRUE,
  dsn = "pssp_prod64"
)
}
\arguments{
\item{catch_area}{Valid catch area code}

\item{start_date}{Start date of search}

\item{end_date}{End date of search}

\item{int_dates}{Returns year, month, day integer columns instead of a date field}

\item{fill_zero}{Fill NA's with zeros.}

\item{fresh}{Boolean. Ensures data is updated}

\item{check_fresh}{Boolean. Checks whether data is current}

\item{dsn}{DSN for the database connection. Default 'pssp_prod64'}
}
\description{
Queries RMIS for release infomation, defaults to tagged releases
}
\examples{
\dontrun{ds_data <- ds_salmon_summary('10', '2021-11-01', '2021-12-04')}
\dontrun{ds_data <- ds_salmon_summary(c('11','10'), '2021-11-01', '2021-12-04')}
}
\seealso{
Other Dockside Creel Surveys: 
\code{\link{ds_bio_pull}()},
\code{\link{ds_encounters}()},
\code{\link{ds_filter_insample}()},
\code{\link{ds_filter_insample_aerial}()},
\code{\link{ds_filter_insample_murthy}()},
\code{\link{ds_historical_cpue_chart}()},
\code{\link{ds_pull_sampling_effort}()},
\code{\link{ds_releases_pull}()},
\code{\link{ds_sampling_effort_sampler_summary}()},
\code{\link{ds_sampling_effort_summary}()},
\code{\link{identify_in_sample}()}
}
\concept{Dockside Creel Surveys}
