% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockside.R
\name{ds_sampling_effort_summary}
\alias{ds_sampling_effort_summary}
\title{Sampling effort summary}
\usage{
ds_sampling_effort_summary(marine_area, start_date, end_date)
}
\arguments{
\item{marine_area}{Valid marine area code}

\item{start_date}{Start date of search}

\item{end_date}{End date of search}
}
\description{
Gets effort data and summarizes by survey trips and survey hours
}
\examples{
\dontrun{df <- ds_sampling_effort_summary('10', '2021-01-01', '2021-12-31')}
}
\seealso{
Other Dockside Creel Surveys: 
\code{\link{ds_bio_pull}()},
\code{\link{ds_encounters}()},
\code{\link{ds_filter_insample}()},
\code{\link{ds_filter_insample_aerial}()},
\code{\link{ds_filter_insample_murthy}()},
\code{\link{ds_historical_cpue_chart}()},
\code{\link{ds_pull_sampling_effort}()},
\code{\link{ds_releases_pull}()},
\code{\link{ds_salmon_summary}()},
\code{\link{ds_sampling_effort_sampler_summary}()},
\code{\link{identify_in_sample}()}
}
\concept{Dockside Creel Surveys}
