% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_sample_filtering.R
\name{identify_in_sample}
\alias{identify_in_sample}
\title{In-Sample idenifier}
\usage{
identify_in_sample(
  .data,
  location_code_column = "location_code",
  design = c("murthy", "aerial"),
  named = FALSE
)
}
\arguments{
\item{.data}{Dataframe}

\item{location_code_column}{The column in dataframe where the location codes are.}

\item{design}{Either murthy or aerial}

\item{named}{Boolen. Returns a column of in/out of sample.}
}
\description{
Identifys in-sample locations by adding in_sampling column
}
\examples{
\dontrun{df |> identify_in_sample(design = 'murthy')}
}
\seealso{
Other Dockside Creel Surveys: 
\code{\link{ds_bio_pull}()},
\code{\link{ds_encounters}()},
\code{\link{ds_filter_insample}()},
\code{\link{ds_filter_insample_aerial}()},
\code{\link{ds_filter_insample_murthy}()},
\code{\link{ds_historical_cpue_chart}()},
\code{\link{ds_pull_sampling_effort}()},
\code{\link{ds_releases_pull}()},
\code{\link{ds_salmon_summary}()},
\code{\link{ds_sampling_effort_sampler_summary}()},
\code{\link{ds_sampling_effort_summary}()}
}
\concept{Dockside Creel Surveys}
