% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pssp-package.R
\docType{package}
\name{pssp-package}
\alias{pssp}
\alias{pssp-package}
\title{pssp: Package to interface with the WDFW Puget Sound sampling database}
\description{
The pssp package serves as an interface between R and the Puget Sound sampling database. It"s goal is to ease the burden of retrieving and analyzing sport creel information in Puget Sound.
}
\author{
\strong{Maintainer}: Ty Garber \email{tyler.garber@dfw.wa.gov}

}
\keyword{internal}
