% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{pssp_last_update}
\alias{pssp_last_update}
\title{Gets the last time materialized views we updated}
\usage{
pssp_last_update(dsn = "pssp_prod64")
}
\arguments{
\item{dsn}{DSN connection default "pssp_prod64"}
}
\description{
Gets the last time materialized views were updated
}
\examples{
\dontrun{
pssp_last_update()
}
}
