% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{pssp_query}
\alias{pssp_query}
\title{Queries the pssp database}
\usage{
pssp_query(query, dsn = "pssp_prod64")
}
\arguments{
\item{query}{SQL Query}

\item{dsn}{DSN connection default "pssp_prod64"}
}
\description{
Queries the pssp database and returns a tibble
}
\examples{
\dontrun{
data <- pssp_query("SELECT * FROM survey LIMIT 10;")
}
}
