% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{pssp_refresh_data}
\alias{pssp_refresh_data}
\title{Refreseshs datasources}
\usage{
pssp_refresh_data(dsn = "pssp_prod64")
}
\arguments{
\item{dsn}{DSN connection default 'pssp_prod64'}
}
\description{
Refreshes all materialized views
}
\examples{
\dontrun{
pssp_refresh_data()
}
}
