% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testfishing.R
\name{tf_interagency_chisq}
\alias{tf_interagency_chisq}
\title{Interagency Fisher / Chi Square Test}
\usage{
tf_interagency_chisq(.data, significance = 0.05)
}
\arguments{
\item{.data}{Dataframe}

\item{significance}{Significance threshold default .05}
}
\description{
Fisher / Chi Square test between agencies and legal mark counts
}
\examples{
\dontrun{tf_data |> tf_interagency_chisq()}

}
\seealso{
Other Test Fishing: 
\code{\link{tf_cv}()},
\code{\link{tf_descriptive_length}()},
\code{\link{tf_effort_summary_week}()},
\code{\link{tf_hist}()},
\code{\link{tf_interagency_histogram}()},
\code{\link{tf_interagency_lm_summary}()},
\code{\link{tf_interagency_summary}()},
\code{\link{tf_lm_summary}()},
\code{\link{tf_lm_summary_week}()},
\code{\link{tf_map}()},
\code{\link{tf_mortalility}()},
\code{\link{tf_pull}()},
\code{\link{tf_pvalue_history}()},
\code{\link{tf_scatter}()},
\code{\link{tf_summary}()}
}
\concept{Test Fishing}
