% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testfishing.R
\name{tf_interagency_histogram}
\alias{tf_interagency_histogram}
\title{Histogram comparison between agencies by length}
\usage{
tf_interagency_histogram(
  .data,
  x_axis,
  x_label = "Total Length (CM)",
  y_label = "Count",
  title = "Interagency Lengths"
)
}
\arguments{
\item{.data}{Dataframe}

\item{x_axis}{Length column. total_length or fork_length}

\item{x_label}{Label of x axis}

\item{y_label}{Label of y axis}

\item{title}{Title of plot}
}
\description{
Histogram
}
\examples{
\dontrun{tf_data |> tf_interagency_histogram(total_length, agency)}

}
\seealso{
Other Test Fishing: 
\code{\link{tf_cv}()},
\code{\link{tf_descriptive_length}()},
\code{\link{tf_effort_summary_week}()},
\code{\link{tf_hist}()},
\code{\link{tf_interagency_chisq}()},
\code{\link{tf_interagency_lm_summary}()},
\code{\link{tf_interagency_summary}()},
\code{\link{tf_lm_summary}()},
\code{\link{tf_lm_summary_week}()},
\code{\link{tf_map}()},
\code{\link{tf_mortalility}()},
\code{\link{tf_pull}()},
\code{\link{tf_pvalue_history}()},
\code{\link{tf_scatter}()},
\code{\link{tf_summary}()}
}
\concept{Test Fishing}
