% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testfishing.R
\name{tf_interagency_summary}
\alias{tf_interagency_summary}
\title{Iteragency summary}
\usage{
tf_interagency_summary(.data)
}
\arguments{
\item{.data}{Dataframe}
}
\description{
Summary of effort and encounters by agency
}
\examples{
\dontrun{
tf_data |> tf_interagency_summary()
}

}
\seealso{
Other Test Fishing: 
\code{\link{tf_cv}()},
\code{\link{tf_descriptive_length}()},
\code{\link{tf_effort_summary_week}()},
\code{\link{tf_hist}()},
\code{\link{tf_interagency_chisq}()},
\code{\link{tf_interagency_histogram}()},
\code{\link{tf_interagency_lm_summary}()},
\code{\link{tf_lm_summary}()},
\code{\link{tf_lm_summary_week}()},
\code{\link{tf_map}()},
\code{\link{tf_mortalility}()},
\code{\link{tf_pull}()},
\code{\link{tf_pvalue_history}()},
\code{\link{tf_scatter}()},
\code{\link{tf_summary}()}
}
\concept{Test Fishing}
