% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testfishing.R
\name{tf_lm_summary}
\alias{tf_lm_summary}
\title{Summarize by Legal Mark category}
\usage{
tf_lm_summary(.data, size_limit = 22)
}
\arguments{
\item{.data}{Dataframe}

\item{size_limit}{Size limit in inches to be used to find legality. Default 22 inchees}
}
\description{
Creates a summary data with counts by legal mark category
}
\examples{
\dontrun{tf_data |> tf_lm_summary()}

}
\seealso{
Other Test Fishing: 
\code{\link{tf_pull}()}
}
\concept{Test Fishing}
