% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testfishing.R
\name{tf_map}
\alias{tf_map}
\title{Map of test fishing encouners}
\usage{
tf_map(.data, ...)
}
\arguments{
\item{.data}{Dataframe}

\item{...}{add_effort = TRUE, Adds a heat map of boat surveys. start_date, end_date and catch_area must also be supplied}
}
\description{
This produces a leaflet map showing test fishing encounters
}
\examples{
\dontrun{tf_data |> tf_map()}

}
\seealso{
Other Test Fishing: 
\code{\link{tf_cv}()},
\code{\link{tf_descriptive_length}()},
\code{\link{tf_effort_summary_week}()},
\code{\link{tf_hist}()},
\code{\link{tf_interagency_chisq}()},
\code{\link{tf_interagency_histogram}()},
\code{\link{tf_interagency_lm_summary}()},
\code{\link{tf_interagency_summary}()},
\code{\link{tf_lm_summary}()},
\code{\link{tf_lm_summary_week}()},
\code{\link{tf_mortalility}()},
\code{\link{tf_pull}()},
\code{\link{tf_pvalue_history}()},
\code{\link{tf_scatter}()},
\code{\link{tf_summary}()}
}
\concept{Test Fishing}
