% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testfishing.R
\name{tf_pull}
\alias{tf_pull}
\title{Downloads test fishing data from pssp}
\usage{
tf_pull(
  catch_area,
  start_date,
  end_date,
  fresh = F,
  check_fresh = T,
  dsn = "pssp_prod64"
)
}
\arguments{
\item{catch_area}{Valid catch area code}

\item{start_date}{Start date of search}

\item{end_date}{End date of search}

\item{fresh}{Boolean. Ensures data is updated}

\item{check_fresh}{Boolean. Checks whether data is current}

\item{dsn}{DSN for the database connection. Default 'pssp_prod64'}
}
\description{
Queries pssp for test fishing encounters
}
\examples{
\dontrun{tf_data <- tf_pull('10', '2021-11-01', '2021-12-04')}
}
\seealso{
Other Test Fishing: 
\code{\link{tf_lm_summary}()}
}
\concept{Test Fishing}
