% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testfishing.R
\name{tf_pvalue_history}
\alias{tf_pvalue_history}
\title{Line plot of p-values from Chi-square}
\usage{
tf_pvalue_history(.data, alpha = 0.05)
}
\arguments{
\item{.data}{Dataframe}

\item{alpha}{Significance metric to evaluate whether statistically sigificant or not}
}
\description{
Line plot of cummulative p-values yieled from cummulative totals of test fishing encounters
}
\examples{
\dontrun{tf_data |> tf_pvalue_history()}

}
\seealso{
Other Test Fishing: 
\code{\link{tf_cv}()},
\code{\link{tf_descriptive_length}()},
\code{\link{tf_effort_summary_week}()},
\code{\link{tf_hist}()},
\code{\link{tf_interagency_chisq}()},
\code{\link{tf_interagency_histogram}()},
\code{\link{tf_interagency_lm_summary}()},
\code{\link{tf_interagency_summary}()},
\code{\link{tf_lm_summary}()},
\code{\link{tf_lm_summary_week}()},
\code{\link{tf_map}()},
\code{\link{tf_mortalility}()},
\code{\link{tf_pull}()},
\code{\link{tf_scatter}()},
\code{\link{tf_summary}()}
}
\concept{Test Fishing}
