% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testfishing.R
\name{tf_scatter}
\alias{tf_scatter}
\title{Creates a ggplot scatter plot of test fishing lengths by adipose clip}
\usage{
tf_scatter(
  .data,
  x_axis,
  y_axis,
  x_label = "Survey Date",
  y_label = "Total Length (CM)",
  title = "Chinook Test Fishing Lengths"
)
}
\arguments{
\item{.data}{Dataframe}

\item{x_axis}{Variable for x axis of plot}

\item{y_axis}{Variable for y axis of plot}

\item{x_label}{Label of x axis}

\item{y_label}{Label of y axis}

\item{title}{Title of plot}
}
\description{
Queries RMIS for release infomation, defaults to tagged releases
}
\examples{
\dontrun{
tf_data |> tf_scatter(survey_datetime, total_length)
}

}
\seealso{
Other Test Fishing: 
\code{\link{tf_cv}()},
\code{\link{tf_descriptive_length}()},
\code{\link{tf_effort_summary_week}()},
\code{\link{tf_hist}()},
\code{\link{tf_interagency_chisq}()},
\code{\link{tf_interagency_histogram}()},
\code{\link{tf_interagency_lm_summary}()},
\code{\link{tf_interagency_summary}()},
\code{\link{tf_lm_summary}()},
\code{\link{tf_lm_summary_week}()},
\code{\link{tf_map}()},
\code{\link{tf_mortalility}()},
\code{\link{tf_pull}()},
\code{\link{tf_pvalue_history}()},
\code{\link{tf_summary}()}
}
\concept{Test Fishing}
