% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtrs.R
\name{vtr_pull}
\alias{vtr_pull}
\title{Downloads VTR data from pssp}
\usage{
vtr_pull(
  catch_area,
  start_date,
  end_date,
  pretty = TRUE,
  fresh = FALSE,
  check_fresh = TRUE
)
}
\arguments{
\item{catch_area}{Valid catch area code}

\item{start_date}{Start date of search}

\item{end_date}{End date of search}

\item{pretty}{Returns an easily readable dataframe}

\item{fresh}{Boolean. Ensures data is updated}

\item{check_fresh}{Boolean. Checks whether data is current}
}
\value{
Dataframe
}
\description{
Queries RMIS for release infomation, defaults to tagged releases
}
\examples{
\dontrun{vtr_data <- vtr_pull('10', '2021-11-01', '2021-12-04')}
}
\concept{Volantary Trip Reports}
