% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{pssp_query}
\alias{pssp_query}
\title{Queries the pssp database}
\usage{
pssp_query(dsn = "pssp_prod64", query)
}
\arguments{
\item{dsn}{DSN connection default 'pssp_prod64'}

\item{query}{SQL Query}
}
\description{
Queries the pssp database and returns a tibble
}
\examples{
\dontrun{
data <- pssp_query("SELECT * FROM survey LIMIT 10;")
}
}
