% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testfishing.R
\name{tf_lm_summary}
\alias{tf_lm_summary}
\title{Summarize by Legal Mark category}
\usage{
tf_lm_summary(.data, size_limit = 22)
}
\arguments{
\item{.data}{Dataframe}

\item{size_limit}{Size limit in inches to be used to find legality. Default 22 inchees}
}
\description{
Creates a summary data with counts by legal mark category
}
\examples{
\dontrun{
tf_data |> tf_lm_summary()
}

}
\seealso{
Other Test Fishing: 
\code{\link{tf_cv}()},
\code{\link{tf_descriptive_length}()},
\code{\link{tf_effort_summary_week}()},
\code{\link{tf_hist}()},
\code{\link{tf_interagency_chisq}()},
\code{\link{tf_interagency_histogram}()},
\code{\link{tf_interagency_lm_summary}()},
\code{\link{tf_interagency_summary}()},
\code{\link{tf_lm_summary_week}()},
\code{\link{tf_map}()},
\code{\link{tf_mortalility}()},
\code{\link{tf_pull}()},
\code{\link{tf_pvalue_history}()},
\code{\link{tf_scatter}()},
\code{\link{tf_summary}()}
}
\concept{Test Fishing}
